requirements.txt
streamlit==1.54.0
opencv-python==4.13.0.92
mediapipe==0.10.32
Pillow==12.1.1
numpy==2.4.2